iRevert "Merge branch 'master' of github.com:hksix/Eventr-CapstoneProject"
This reverts commit f45bb5fe41270d22efb6798459c723e3f6a6d403, reversing
changes made to 8060118d0715e9af2ed982be0f3e683d91b29e91.

+++ restify-api/migrations/20171003194939-create-event-inventory.js
@@ -27,7 +27,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       ownerid: {
-        type: Sequelize.BIGINT
+        type: Sequelize.INTEGER
       },
       description: {
         type: Sequelize.STRING
